requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
pkg-resources==0.0.0
PyJWT==2.5.0
PyMySQL==1.0.2
pytz==2022.4
sqlparse==0.4.3
